requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
